# backend/src/Api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Skylens API",
    "description": "ADS-B plane-spotter gateway — API for the Skylens app.\n\n---\n\n**Version:** `0.0.0-dev`",
    "version": "v1"
  },
  "paths": {
...
    "/api/client-log": {
      "post": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "ClientLog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Skylens.Api.Endpoints.ApiEndpoints.ClientLogBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/aircraft": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "AircraftList",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Skylens.Api.State.AircraftDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/aircraft/{hex}": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "AircraftDetail",
        "parameters": [
          {
            "name": "hex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skylens.Api.Endpoints.ApiEndpoints.AircraftDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/aircraft/{hex}/route": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "AircraftRoute",
        "parameters": [
          {
            "name": "hex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skylens.Api.Enrichment.FlightRoute"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/aircraft/{hex}/route/cached": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "AircraftRouteCached",
        "parameters": [
          {
            "name": "hex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skylens.Api.Enrichment.FlightRoute"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/area": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "Area",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Skylens.Api.State.AircraftDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vessels": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "VesselList",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Skylens.Api.State.VesselDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vessels/{mmsi}": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "VesselDetail",
        "parameters": [
          {
            "name": "mmsi",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skylens.Api.Endpoints.ApiEndpoints.VesselDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/satellites/{noradId}": {
      "get": {
        "tags": [
          "Skylens.Api, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "SatelliteDetail",
        "parameters": [
          {
            "name": "noradId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skylens.Api.Endpoints.ApiEndpoints.SatelliteDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Skylens.Api.Endpoints.ApiEndpoints.AircraftDetail": {
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/Skylens.Api.State.AircraftDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/Skylens.Api.Enrichment.AircraftMetadata"
          }
        },
        "additionalProperties": false
      },
      "Skylens.Api.Endpoints.ApiEndpoints.ClientLogBatch": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skylens.Api.Endpoints.ApiEndpoints.ClientLogEntry"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Skylens.Api.Endpoints.ApiEndpoints.ClientLogEntry": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "nullable": true
          },
          "endpoint": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "edgeMarkerPresent": {
            "type": "boolean"
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Skylens.Api.Endpoints.ApiEndpoints.SatelliteDetail": {
        "type": "object",
        "properties": {
          "satellite": {
            "$ref": "#/components/schemas/Skylens.Api.Enrichment.SatelliteDto"
          },
          "transmitters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skylens.Api.Enrichment.SatelliteTransmitterDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Skylens.Api.Endpoints.ApiEndpoints.VesselDetail": {
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/Skylens.Api.State.VesselDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/Skylens.Api.Enrichment.VesselMetadata"
          }
        },
        "additionalProperties": false
      },
...
      "Skylens.Api.Enrichment.AircraftMetadata": {
        "required": [
          "hex"
        ],
        "type": "object",
        "properties": {
          "hex": {
            "type": "string",
            "nullable": true
          },
          "registration": {
            "type": "string",
            "nullable": true
          },
          "typeCode": {
            "type": "string",
            "nullable": true
          },
          "typeName": {
            "type": "string",
            "nullable": true
          },
          "operator": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Skylens.Api.Enrichment.FlightRoute": {
        "required": [
          "ident"
        ],
        "type": "object",
        "properties": {
          "ident": {
            "type": "string",
            "nullable": true
          },
          "originIcao": {
            "type": "string",
            "nullable": true
          },
          "originName": {
            "type": "string",
            "nullable": true
          },
          "destinationIcao": {
            "type": "string",
            "nullable": true
          },
          "destinationName": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Skylens.Api.State.AircraftDto": {
        "required": [
          "hex"
        ],
        "type": "object",
        "properties": {
          "hex": {
            "type": "string",
            "nullable": true
          },
          "flight": {
            "type": "string",
            "nullable": true
          },
          "fl": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lat": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lon": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "alt": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gs": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "trk": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "vr": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "seen": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cat": {
            "type": "string",
            "nullable": true
          },
          "src": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Skylens.Api.State.VesselDto": {
        "required": [
          "kind",
          "mmsi"
        ],
        "type": "object",
        "properties": {
          "mmsi": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "kind": {
            "type": "string",
            "nullable": true
          },
          "lat": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lon": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "sog": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cog": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "hdg": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "shipType": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "navStatus": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aidType": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "flag": {
            "type": "string",
            "nullable": true
          },
          "seen": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "src": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}